moved agent-related code and rng-related code to separate files
