+++ src/CMakeLists.txt
@@ -1,5 +1,5 @@
 # Define source files not defining "main" as a static library for linking
-add_library(game STATIC game.cpp)
+add_library(game STATIC game.cpp agent.cpp rng.cpp)
 
 # Link compiler_flags (defined at top level)
 target_link_libraries(game PUBLIC compiler_flags)
