+++ src/game.hpp
@@ -11,6 +11,8 @@
 #include <span>
 #include <vector>
 
+#include "agent.hpp"
+
 namespace GameConstants {
     static constexpr size_t MIN_PLAYERS = 2;
     static constexpr size_t MAX_PLAYERS = 5;
@@ -27,12 +29,6 @@ namespace GameConstants {
 enum class ActionType {
     First,
     Second
-};
-
-// TODO: move to policy file
-enum class PolicyType {
-    Random,
-    Greedy
 };
 
 enum class Color { 
@@ -106,29 +102,6 @@ struct State {
         num_locks(0),
         is_terminal(false)
         {};
-};
-
-class Agent {
-public:
-    Agent(size_t position) : m_position(position) {};
-
-    virtual std::optional<size_t> make_move(std::span<const Move> moves, const State& state) const = 0;
-protected:
-    size_t m_position;
-};
-
-class RandomAgent : public Agent {
-public:
-    RandomAgent(size_t position) : Agent(position) {};
-    std::optional<size_t> make_move(std::span<const Move> moves, const State& state) const override;
-};
-
-class GreedyAgent : public Agent {
-public:
-    GreedyAgent(size_t position, int max_skips) : Agent(position), m_max_skips(max_skips) {};
-    std::optional<size_t> make_move(std::span<const Move> moves, const State& state) const override;
-protected:
-    int m_max_skips;
 };
 
 struct GameData {
