+++ src/rng.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <random>
+
+extern std::mt19937_64& rng();
